requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Utilities
azure-identity
azure-keyvault-secrets
azure-storage-blob
msrestazure
python-dotenv
winshell

# Data 
pandas
pyarrow
openpyxl
pyodbc

# Web
requests
jsonschema
Flask
fastapi
uvicorn

# Lowlevel
orjson


